requirements.txt
# Main application and ML libraries
gradio==3.50.2
albumentations==1.3.0
basicsr==1.4.2
einops==0.4.1
facexlib==0.3.0
gfpgan==1.3.8
imageio==2.9.0
imageio-ffmpeg==0.4.2
lmdb==1.3.0
lpips==0.1.4
numpy==1.23.1
opencv-python==4.6.0.66
Pillow==9.2.0
pyyaml==6.0
realesrgan==0.3.0
scikit-image==0.19.3
scipy==1.9.1
tqdm==4.64.1
yapf==0.32.0
addict==2.4.0
future==0.18.2
pytorch-lightning==1.4.2
requests==2.28.1
torch-fidelity==0.3.0
torchmetrics==0.7.0
voluptuous==0.13.1
transformers==4.30.2
tokenizers==0.13.3
kornia
open_clip_torch==2.7.0
xformers==0.0.16